rewrote b64 decoder to work with older browsers
Previous function only worked for IE10+ and other browsers with native window.atob() functions. The pass-through case creating a new Buffer object assumed the code was running on Node.js, but our code runs in the Angular front-end and so would fail on older browsers. The new function takes advantage of a client-side script by Tyler Akins as noted in the comments.